add comment markdown support
