+++ src/Components/markdown-content/style.css
@@ -1,0 +1,23 @@
+.cac_markdown {
+  width:100%;
+}
+
+.cac_markdown img{
+  max-width: 100%;
+} 
+
+.cac_markdown pre {
+  border: 1.5px solid rgb(173, 173, 173);
+  border-radius: 3px;
+
+  background-color: rgb(250, 250, 250) !important;
+  font-size: 1.2em !important;
+  padding-top: 0.5rem !important;
+  padding-left: 1rem !important;
+  padding-bottom: .5rem !important;
+  margin-left: 2rem !important;
+  margin-right: 2rem !important;
+}
+.cac_markdown pre .token {
+  background: transparent !important;
+}
